\name{run_app}
\alias{run_app}
\title{Run the MatriSpace Shiny Application}
\usage{
run_app(host = "127.0.0.1", port = NULL, ...)
}
\arguments{
\item{host}{The host address. Defaults to "127.0.0.1".}
\item{port}{The port number. Defaults to NULL (random available port).}
\item{...}{Additional arguments passed to golem options.}
}
\description{
Launches the MatriSpace Shiny application for offline spatial
transcriptomics matrisome analysis.
}
